[Feature][flink-k8s] Correct package name
