+++ streampark-flink/streampark-flink-kubernetes/src/main/java/org/apache/streampark/flink/kubernetes/PodTemplatePool.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.flink.kubernetes;
+package org.apache.streampark.flink.kubernetes;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.kubernetes.model.K8sPodTemplates;
+import org.apache.streampark.flink.kubernetes.model.K8sPodTemplates;
 
 import java.io.File;
 import java.io.IOException;
