+++ streampark-flink/streampark-flink-kubernetes/src/main/java/org/apache/streampark/flink/kubernetes/enums/FlinkK8sExecuteMode.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.kubernetes.enums;
+package org.apache.streampark.flink.kubernetes.enums;
 
 import org.apache.streampark.common.enums.ExecutionMode;
