+++ streampark-flink/streampark-flink-kubernetes/src/main/java/org/apache/streampark/flink/kubernetes/event/FlinkJobCheckpointChangeEvent.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.flink.kubernetes.event;
+package org.apache.streampark.flink.kubernetes.event;
 
-import org.apache.flink.kubernetes.model.CheckpointCV;
-import org.apache.flink.kubernetes.model.TrackId;
+import org.apache.streampark.flink.kubernetes.model.CheckpointCV;
+import org.apache.streampark.flink.kubernetes.model.TrackId;
 
 /** held internally by K8sFlinkMonitor. */
 public class FlinkJobCheckpointChangeEvent implements BuildInEvent {
